[WIP][REFACTOR] Moving ControllerActions + its subfunctions to Requestforms
